// frontend/src/types.ts
export interface ProblemListItem {
  id: number
  title: string
  difficulty: number
  tags: string[]
  accepted_count: number
  submission_count: number
  created_at: string
  updated_at: string
  my_status?: 'accepted' | 'wrong_answer'
  // 管理员列表专属
  type?: ProblemType
  status?: ProblemStatus
  testcase_count?: number
}

export type ProblemStatus = 'draft' | 'published' | 'disabled'
export type ProblemType = 'standard' | 'spj' | 'interactive' | 'output_only'

export interface Sample {
  input: string
  output: string
  note: string
}

export interface ProblemDetail extends ProblemListItem {
  statement: string
  input_format: string
  output_format: string
  hint: string
  samples: Sample[]
  time_limit_ms: number
  memory_limit_kb: number
  type: ProblemType
  testcase_scores: number[]
  submission_limit: number
  status: ProblemStatus
  is_favorite?: boolean
  // 管理员专属
  spj_source?: string
  interactor_source?: string
  testcase_count?: number
}
// ...
export interface Language {
  key: string
  name: string
  version: string
  monaco: string
  supports_optimize: boolean
}